Optimistic minimal draft implementation
Implementing an optimistic minimal draft, that is able to generate a simple jetbrains/exposed table based on a simple sql create table statement.
It's possible to resolve varchar fields, and set a single primary key field.

The sql file is first split into tokens, and the tokens are parsed into an in memory representation of the sql file at hand. The in memory representation get then translated into an instruction set, used by a stack like vm to generate kotlin code for the exposed table.

The code is in a rough and drafty state.