Fixing an issue with NOT NULL evaluation
Parsing an sql statement that contains a NOT NULL constraint and storing it in a statement type, the `isNullable` method used an `any` check on the constrains collection, which resulted in the wrong value. Now a simple contains is used to check if a statement can be null or not.