Creating lexer and parser for oracle dialect
Adding tokens for the different keywords which are neccessary in oracle
and other dialects.

Creating factory methods for getting a oracle dialect specific lexer and
parser. For later extension other dialect specific lexers and parsers
can be added.

Changed common lexer for disassemble the sql in a parsable token list.

Created a oracle lexer which contains the tokenizing for the unique and
constraint.

Created oracle parser which takes the list of tokens and creates an
object tree which can be interpreted. The parser differs between the
table creation, the column declarations with column constraints and the
table contraints.
The parser creates a table object with the columns and constraints.

The lexing and parsing algorithm is tested via Unit tests.