Feature/oracle generator (#2)
* Creating lexer and parser for oracle dialect

Adding tokens for the different keywords which are neccessary in oracle
and other dialects.

Creating factory methods for getting a oracle dialect specific lexer and
parser. For later extension other dialect specific lexers and parsers
can be added.

Changed common lexer for disassemble the sql in a parsable token list.

Created a oracle lexer which contains the tokenizing for the unique and
constraint.

Created oracle parser which takes the list of tokens and creates an
object tree which can be interpreted. The parser differs between the
table creation, the column declarations with column constraints and the
table contraints.
The parser creates a table object with the columns and constraints.

The lexing and parsing algorithm is tested via Unit tests.

* Adding instruction generating. In addition adding new instructions type for foreign keys and unique contraints. Moreover adding the default constraint to the columnStatement instruction.

Changed interpreter which prints the exposed code on the console. We initially ignore the instructions for foreign key and unique, because it has no additional benefit in the software. These constraints will only evaluated in the database.

Added Copyright to every class.

* Create gradle.yml

* Creating lexer and parser for oracle dialect

Adding tokens for the different keywords which are neccessary in oracle
and other dialects.

Creating factory methods for getting a oracle dialect specific lexer and
parser. For later extension other dialect specific lexers and parsers
can be added.

Changed common lexer for disassemble the sql in a parsable token list.

Created a oracle lexer which contains the tokenizing for the unique and
constraint.

Created oracle parser which takes the list of tokens and creates an
object tree which can be interpreted. The parser differs between the
table creation, the column declarations with column constraints and the
table contraints.
The parser creates a table object with the columns and constraints.

The lexing and parsing algorithm is tested via Unit tests.

* Adding instruction generating. In addition adding new instructions type for foreign keys and unique contraints. Moreover adding the default constraint to the columnStatement instruction.

Changed interpreter which prints the exposed code on the console. We initially ignore the instructions for foreign key and unique, because it has no additional benefit in the software. These constraints will only evaluated in the database.

Added Copyright to every class.

* Changed chmod -x to make gradlew executable on linux build environment

* Fixed testcase

Co-authored-by: cmanig1 <christoph.manig@isb-ag.de>