Treating default values not as int
Fixing an issue where we assumed that a default value in sql is always an int. This is wrong, as a default value could be anything. As we actually don't care about the type behind the default value, as we're only using it for code generation, we now use simply string to represent the actual value.