Fixing casing for the object name in kotlin
For all properties of the exposed table object, we want to start with a lowercase char. This mechanic found its way into the object naming, but that's wrong, as objects should always start with an upper case character in kotlin.
This commit fixes the issue by extracting the first character and making it titlecase, adding the rest of the string at the end