Fixing compilation error in test
Due to the switch of the used type for default values (int to string, as we don't care about the actual value and its type), a test failed with a compilation error, as the test case was passing an int value as default value, but should've been a string. This is fixed in this commit.